# Laboratorios/compra-entradas/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Compra de Entradas",
        "description": "Documentación de la API para gestión de tickets y pagos",
        "contact": {
            "email": "soporte@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
        "/api/v1/tickets": {
            "description": "Operaciones sobre compra de tickets",
            "get": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Listar mis tickets",
                "operationId": "b406702bcd80642f4b5c83c1bf20105b",
                "responses": {
                    "200": {
                        "description": "Listado de tickets",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Ticket"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Comprar boletos (requiere JWT)",
                "operationId": "80fbef860b264eed3bc3ccd90d8ddc8b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "event_id",
                                    "quantity",
                                    "price"
                                ],
                                "properties": {
                                    "event_id": {
                                        "type": "integer"
                                    },
                                    "quantity": {
                                        "type": "integer"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ticket comprado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ticket"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/events": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Listar eventos disponibles",
                "operationId": "b2e55feb6b984fe8bede8d15c4028b7d",
                "responses": {
                    "200": {
                        "description": "Lista de eventos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {}
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "Error al comunicarse con el servicio de Eventos"
                    }
                }
            }
        },
        "/api/v1/events/{id}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Obtener detalle de un evento",
                "operationId": "cab1311cf1b9c45ea4813e7b5d0fd908",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del evento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle del evento",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Evento no encontrado"
                    },
                    "502": {
                        "description": "Error al comunicarse con el servicio de Eventos"
                    }
                }
            }
        },
        "/api/v1/pagar/{ticket_id}": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Confirmar pago de un ticket (requiere JWT)",
                "operationId": "4114ea2bbadc49e06242c58d4beb50a0",
                "parameters": [
                    {
                        "name": "ticket_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pago confirmado"
                    },
                    "404": {
                        "description": "Pago no encontrado"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/tickets/{ticket}": {
            "get": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Mostrar detalle de un ticket",
                "operationId": "3ab69850c5fa8c61fde14f244f76bdea",
                "parameters": [
                    {
                        "name": "ticket",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle del ticket",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ticket"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ticket no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Actualizar un ticket",
                "operationId": "ac358141cbf679d00d6e7d6a3e44e232",
                "parameters": [
                    {
                        "name": "ticket",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "quantity": {
                                        "type": "integer"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ticket actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ticket"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ticket no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Eliminar un ticket",
                "operationId": "334f4d33891294369023fc6f0ece7f2c",
                "parameters": [
                    {
                        "name": "ticket",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Ticket eliminado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Payment": {
                "title": "Payment",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "ticket_id": {
                        "type": "integer",
                        "example": 10
                    },
                    "status": {
                        "type": "string",
                        "example": "confirmed"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Ticket": {
                "title": "Ticket",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "event_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 2
                    },
                    "total_price": {
                        "type": "number",
                        "format": "float",
                        "example": 100
                    },
                    "status": {
                        "type": "string",
                        "example": "pendiente"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Introduce \"Bearer {token}\""
            }
        }
    },
    "tags": [
        {
            "name": "Events",
            "description": "Operaciones para obtener eventos desde el microservicio de Eventos"
        },
        {
            "name": "Payments",
            "description": "Operaciones para pagos de tickets"
        },
        {
            "name": "Tickets",
            "description": "Tickets"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ]
}